+++ s2gos-common/src/s2gos_common/process/cli/__init__.py
@@ -1,6 +1,0 @@
-#  Copyright (c) 2025 by ESA DTE-S2GOS team and contributors
-#  Permissions are hereby granted under the terms of the Apache 2.0 License:
-#  https://opensource.org/license/apache-2-0.
-
-# No exports here. To avoid unnecessary imports, CLI apps should pick
-# only what they need from dedicated submodules.