+++ CHANGES.md
@@ -1,3 +1,10 @@
+## Changes in version 0.0.6 (in development)
+
+- Implementation code moved to [Eozilla](https://github.com/eo-tools/eozilla) (#39):
+  - Package `s2gos-client` now uses the Eozilla package `cuiman`.
+  - Package `s2gos-server` now uses the Eozilla package `wraptile`.
+  - Package `s2gos-app-ex` now uses the Eozilla package `procodile`.
+
 ## Changes in version 0.0.5
 
 - In `s2gos_client`: